Add preprocessor output to hash computation
This commit reverses some of the changes introduced by #24 with the goal to
include the output of Clang preprocessing in the hash computation.

The handling of the compiler arguments is modified to include the `-P` argument
along with `-E`. This removes from the preprocessor output the linemarkers (i.e.
line starting with `#`), which in turn makes it possible to directly use that
output in the hash computation without parsing it line by line (which was one of
the issue in #24).

The hash computation itself is modified as follows:

- Execute the preprocessor command and add the output to the hash computation.
- Remove the file modification time from the hash computation.

These changes are designed to make the hash depends on any changes in the
translation unit being analyzed.

Closes #26