Add save output for server cache storage
The output of clang tidy is saved on the server in files that have the name of
the calculated hash. For data that was not saved before the reboot, a separate
cleaning function was made - "_do_cleanup_invalid". To control memory on the
server, the "--max-cache-size" or "-M" flag was added. Using this flag, you can
configure the approximate size of disk space allowed for use. Also added is the
removal of files that have not been used for the longest time, which is
reflected in the "_do_cleanup_least_recently_used" function.

The response for the client is either a 200 response; if the response has a
different status, then there is no data or the request is not valid.