Fix the 'clang-tidy' wrapper script
In 05f3db7c716eba8cbbdbba880fc1aedb84a67654 (move to src-layout), the
file permission of clang-tidy-cache.py is 644; before that the
clang-tidy-cache script had 755. This caused the 'clang-tidy' script to
fail when invoking clang-tidy-cache.py.

Run the script via `/usr/bin/env python3` (equivalent to the shbang in
clang-tidy-cache.py) to fix this.